requirements.txt
flask==3.0.0
marker-pdf>=1.0.0
torch>=2.0.0
gunicorn==21.2.0
requests>=2.31.0
